Updated NavigationCamera to allow injection of a custom following frame calculation strategies.
- Introduced FollowingCameraFramingStrategy, an interface that defines a strategy used to calculate points to be framed for the FOLLOWING camera mode.
- Moved ViewportDataSourceProcessor.getPointsToFrameOnCurrentStep and ViewportDataSourceProcessor.getPointsToFrameAfterCurrentManeuver into MapboxFollowingCameraFramingStrategy.
- Introduced FollowingFrameOptions.framingStrategy that allows injection of custom FollowingCameraFramingStrategy.
- Updated MapboxNavigationViewportDataSource to use ollowingFrameOptions.framingStrategy
- Added CustomFollowingCameraFramingStrategy to examples app to showcase introduced customization
- regenerated libnavui-maps metalava file
