+++ examples/src/main/java/com/mapbox/navigation/examples/core/camera/AnimationType.kt
@@ -7,5 +7,6 @@ enum class AnimationType {
     FastFollowing,
     ToPOI,
     LookAtPOIWhenFollowing,
-    RemoveRoute
+    RemoveRoute,
+    CustomFollowing
 }
